Rate bug fix on first launch

+++ CurrencyConverter/Model/ExchangeRateData.swift
@@ -9,5 +9,5 @@
 import Foundation
 
 struct ExchangeRateData: Codable {
-    let rates: [String: Float]
+    let rates: [String: Double]
 }
